pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolov11-open-source"
version = "0.1.0"
description = "A clean, modular, open-source implementation of YOLOv11"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Open Source Community"},
]
maintainers = [
    {name = "Open Source Community"},
]
keywords = [
    "yolo",
    "yolov11",
    "object-detection",
    "instance-segmentation",
    "pose-estimation",
    "computer-vision",
    "deep-learning",
    "pytorch",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "numpy>=1.21.0",
    "Pillow>=9.0.0",
    "matplotlib>=3.5.0",
    "tqdm>=4.64.0",
    "PyYAML>=6.0",
    "pandas>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]
export = [
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
]
all = [
    "yolov11-open-source[dev,export]",
    "albumentations>=1.3.0",
    "tensorboard>=2.10.0",
    "thop>=0.1.1",
]

[project.urls]
Homepage = "https://github.com/your-org/yolov11-open-source"
Documentation = "https://github.com/your-org/yolov11-open-source#readme"
Repository = "https://github.com/your-org/yolov11-open-source"
Issues = "https://github.com/your-org/yolov11-open-source/issues"

[project.scripts]
yolo-train = "scripts.train:main"
yolo-infer = "scripts.inference:main"
yolo-transfer = "scripts.transfer_weights:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["yolo*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["yolo"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

